+++ api/_utils.ts
@@ -891,6 +891,12 @@ export function positiveIntStr() {
   });
 }
 
+export function hexString() {
+  return define<string>("hexString", (value) => {
+    return utils.isHexString(value as string);
+  });
+}
+
 export function positiveFloatStr(maxValue?: number) {
   return define<string>("positiveFloatStr", (value) => {
     return Number(value) >= 0 && (maxValue ? Number(value) <= maxValue : true);
